[refactor] 인증관련 코드 멘토님 피드백 반영
