+++ src/main/java/com/back/domain/auth/dto/request/LoginRequest.java
@@ -12,8 +12,6 @@ public record LoginRequest(
         @NotBlank(message = "비밀번호는 필수입니다.")
         String password,
 
-        Role selectedRole,
-
-        String loginIp
+        Role selectedRole
 ) {
 }
