+++ src/main/java/com/back/domain/auth/entity/UserToken.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 
 @Getter
 @Entity
-@Table(name = "UsersTokens")
+@Table(name = "user_tokens")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Builder
