+++ src/main/java/com/back/domain/user/entity/User.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 @Getter
 @Entity
-@Table(name = "Users")
+@Table(name = "users")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class User extends BaseEntity {
 
@@ -58,11 +58,8 @@ public class User extends BaseEntity {
 
     private LocalDateTime termsAgreedAt;
 
-    private String agreementIp;
-
     // 정적 팩토리 메서드 - 로컬 회원가입
-    public static User createLocalUser(String email, String password, String name, String phone,
-                                       String agreementIp) {
+    public static User createLocalUser(String email, String password, String name, String phone) {
         User user = new User();
         user.email = email;
         user.password = password;
@@ -77,7 +74,6 @@ public class User extends BaseEntity {
         user.privacyRequiredAgreed = true;  // 회원가입 시 필수 동의
         user.marketingAgreed = false;       // 선택 동의 (기본값)
         user.termsAgreedAt = LocalDateTime.now();  // 현재 시간
-        user.agreementIp = agreementIp;    // 외부에서 받음
         return user;
     }
 
@@ -90,6 +86,37 @@ public class User extends BaseEntity {
     public boolean canLoginAs(Role targetRole) {
         return role.canLoginAs(targetRole);
     }
+
+
+    /**
+     * Tell, Don't Ask 메서드들
+     */
+
+    // 관리자인지 확인
+    public boolean isAdmin() {
+        return Role.ADMIN.equals(this.role);
+    }
+
+    // 아티스트인지 확인
+    public boolean isArtist() {
+        return Role.ARTIST.equals(this.role) && Boolean.TRUE.equals(this.isArtistVerified);
+    }
+
+    // 돈이 충분한지 확인
+    public boolean hasEnoughMoney(int amount) {
+        return this.money >= amount;
+    }
+
+    // 본인인지 확인 (ID 비교)
+    public boolean isSameUser(Long userId) {
+        return userId != null && this.getId().equals(userId);
+    }
+
+    // 본인인지 확인 (User 객체 비교)
+    public boolean isSameUser(User user) {
+        return user != null && this.getId().equals(user.getId());
+    }
+
 
     // TODO: 정적 팩토리 메서드 - 소셜 로그인 구현
 }
