--- src/main/java/com/back/global/util/IpUtils.java
@@ -1,47 +1,0 @@
-package com.back.global.util;
-
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.stereotype.Component;
-
-/**
- * 클라이언트 IP 주소를 추출하는 유틸리티 클래스
- *
- * 프록시, 로드밸런서 환경을 고려하여 실제 클라이언트 IP를 추출합니다.
- * X-Forwarded-For, X-Real-IP 등 다양한 헤더를 순차적으로 확인하여
- * 가장 적절한 클라이언트 IP 주소를 반환합니다.
- *
- * @Author ahdeka
- * @Since 2025-09-24
- */
-@Component
-public class IpUtils {
-
-    public static String getClientIp(HttpServletRequest request) {
-        String ip = request.getHeader("X-Forwarded-For");
-        if (ip == null || ip.isEmpty() || "unknown".equalsIgnoreCase(ip)) {
-            ip = request.getHeader("X-Real-IP");
-        }
-        if (ip == null || ip.isEmpty() || "unknown".equalsIgnoreCase(ip)) {
-            ip = request.getHeader("Proxy-Client-IP");
-        }
-        if (ip == null || ip.isEmpty() || "unknown".equalsIgnoreCase(ip)) {
-            ip = request.getHeader("WL-Proxy-CLIENT-IP");
-        }
-        if (ip == null || ip.isEmpty() || "unknown".equalsIgnoreCase(ip)) {
-            ip = request.getHeader("HTTP_CLIENT_IP");
-        }
-        if (ip == null || ip.isEmpty() || "unknown".equalsIgnoreCase(ip)) {
-            ip = request.getHeader("HTTP_X_FORWARDED_FOR");
-        }
-        if (ip == null || ip.isEmpty() || "unknown".equalsIgnoreCase(ip)) {
-            ip = request.getRemoteAddr();
-        }
-
-        // 로드밸런서를 통과한 경우 첫 번째 IP 반환
-        if (ip != null && ip.contains(",")) {
-            ip = ip.split(",")[0].trim();
-        }
-
-        return ip;
-    }
-}
